refactor: remove dead code and fix sensor caching
- Remove dead timed-spin path from BluetoothCommandParser
- Remove unused clear_mask bitmask from DirectionalScan
- Remove unused SafetyManager::set_emergency_stop and reset_min_voltage
- Remove unused InputWatchdog::_motion_active field
- Simplify debug toggle in BluetoothSystemCommands
- Fix Config.h comment on AUTO_SPIN_* constants
- Cache Proximity structs in ObstacleDetection to stop ping storm
- Share one battery sample per loop between SafetyManager and BatteryVoltage
- Cast float args to double at snprintf sites

+++ src/safety/motion_policy.cpp
@@ -102,7 +102,7 @@ MotionCommand apply_safety(MotionCommand cmd) {
     cmd.rotate  *= scale;
 
 	DBG_PRINT(Debug::Ch::SAFETY, "SCALE=%.2f F=%.2f S=%.2f R=%.2f",
-			  scale, cmd.forward, cmd.strafe, cmd.rotate);
+			  (double)scale, (double)cmd.forward, (double)cmd.strafe, (double)cmd.rotate);
 
     return cmd;
 }
